+++ doc/src/modules/physics/index.rst
@@ -19,11 +19,10 @@ Contents
     sho.rst
     secondquant.rst
     wigner.rst
-    units.rst
     hep/index.rst
     vector/index.rst
     mechanics/index.rst
     quantum/index.rst
     optics/index.rst
-    unitsystems/index.rst
+    units/index.rst
     continuum_mechanics/index.rst
