+++ setup.py
@@ -107,8 +107,8 @@ modules = [
     'sympy.physics.mechanics',
     'sympy.physics.optics',
     'sympy.physics.quantum',
-    'sympy.physics.unitsystems',
-    'sympy.physics.unitsystems.systems',
+    'sympy.physics.units',
+    'sympy.physics.units.systems',
     'sympy.physics.vector',
     'sympy.plotting',
     'sympy.plotting.intervalmath',
@@ -291,7 +291,7 @@ tests = [
     'sympy.physics.optics.tests',
     'sympy.physics.quantum.tests',
     'sympy.physics.tests',
-    'sympy.physics.unitsystems.tests',
+    'sympy.physics.units.tests',
     'sympy.physics.vector.tests',
     'sympy.plotting.intervalmath.tests',
     'sympy.plotting.pygletplot.tests',
