+++ sympy/physics/units/systems/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from sympy.physics.units.systems.mks import _mks_dim, MKS
+from sympy.physics.units.systems.mksa import _mksa_dim, MKSA
+from sympy.physics.units.systems.natural import _natural_dim, natural
