+++ sympy/physics/unitsystems.py
@@ -1,0 +1,7 @@
+"""
+Deprecated: use ``sympy.physics.units``
+"""
+
+# DEPRECATED: use `units`
+
+exec("from .units import *")
