--- sympy/physics/unitsystems/systems/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from sympy.physics.unitsystems.systems.mks import _mks_dim, MKS
-from sympy.physics.unitsystems.systems.mksa import _mksa_dim, MKSA
-from sympy.physics.unitsystems.systems.natural import _natural_dim, natural
